refactor(settings): split settings dialog part files into modules
