+++ lib/src/features/settings/presentation/panes/general_pane.dart
@@ -1,7 +1,26 @@
-part of 'settings_dialog.dart';
+import 'package:alera/src/app/providers.dart';
+import 'package:alera/src/app/theme/alera_tokens.dart';
+import 'package:alera/src/design_system/forms/alera_setting_row.dart';
+import 'package:alera/src/design_system/forms/alera_text_field.dart';
+import 'package:alera/src/design_system/icons/alera_icons.dart';
+import 'package:alera/src/design_system/layout/alera_settings_group.dart';
+import 'package:alera/src/design_system/surfaces/alera_panel.dart';
+import 'package:alera/src/features/agent_status/domain/agent_status.dart';
+import 'package:alera/src/features/settings/domain/alera_settings.dart';
+import 'package:alera/src/features/settings/presentation/panes/agents_cli_skill_control.dart';
+import 'package:alera/src/features/settings/presentation/panes/application_support_section.dart';
+import 'package:alera/src/features/settings/presentation/rows/settings_rows.dart';
+import 'package:alera/src/features/updater/presentation/update_settings_section.dart';
+import 'package:file_selector/file_selector.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-class _GeneralSettingsPane extends ConsumerWidget {
-  const _GeneralSettingsPane({required this.general, required this.agents});
+class GeneralSettingsPane extends ConsumerWidget {
+  const GeneralSettingsPane({
+    super.key,
+    required this.general,
+    required this.agents,
+  });
 
   final GeneralSettings general;
   final AgentSettings agents;
@@ -23,12 +42,12 @@ class _GeneralSettingsPane extends ConsumerWidget {
           ],
         ),
         const SizedBox(height: AleraTokens.space16),
-        _SettingsGroup(
+        AleraSettingsGroup(
           title: 'Safety',
           description:
               'Confirmation prompts for destructive workspace actions.',
           children: <Widget>[
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'Confirm Project Removal',
               description:
                   'Ask before unregistering a project and deleting its workspace metadata.',
@@ -37,7 +56,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
                   .read(settingsControllerProvider.notifier)
                   .setConfirmProjectRemoval(value),
             ),
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'Confirm Workspace Removal',
               description:
                   'Ask before removing a linked workspace and deleting its branch.',
@@ -49,7 +68,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
           ],
         ),
         const SizedBox(height: AleraTokens.space16),
-        _SettingsGroup(
+        AleraSettingsGroup(
           title: 'Agent status',
           description: 'Managed hooks let terminal tabs show agent state.',
           children: <Widget>[
@@ -58,9 +77,9 @@ class _GeneralSettingsPane extends ConsumerWidget {
               description:
                   'Install The Codex Skill That Teaches Agents To Use The Alera CLI.',
               controlWidth: 280,
-              child: _AleraCliSkillControl(),
+              child: AleraCliSkillControl(),
             ),
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'Codex Hooks',
               description:
                   'Use an Alera-managed Codex runtime home with status hooks.',
@@ -69,7 +88,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
                   .read(settingsControllerProvider.notifier)
                   .setAgentStatusHookEnabled(AgentType.codex, value),
             ),
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'Claude Code Hooks',
               description:
                   'Use an Alera-managed Claude Code config with status hooks.',
@@ -78,7 +97,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
                   .read(settingsControllerProvider.notifier)
                   .setAgentStatusHookEnabled(AgentType.claude, value),
             ),
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'GitHub Copilot Hooks',
               description: 'Use an Alera-managed GitHub Copilot home overlay.',
               value: agents.agentStatusHooks.copilot,
@@ -86,7 +105,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
                   .read(settingsControllerProvider.notifier)
                   .setAgentStatusHookEnabled(AgentType.copilot, value),
             ),
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'Cursor Hooks',
               description: 'Use an Alera-managed Cursor Agent plugin wrapper.',
               value: agents.agentStatusHooks.cursor,
@@ -94,7 +113,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
                   .read(settingsControllerProvider.notifier)
                   .setAgentStatusHookEnabled(AgentType.cursor, value),
             ),
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'Antigravity Hooks',
               description:
                   'Install Alera-managed Antigravity hooks for the agy CLI. Disable to remove only Alera-managed hook entries.',
@@ -103,7 +122,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
                   .read(settingsControllerProvider.notifier)
                   .setAgentStatusHookEnabled(AgentType.agy, value),
             ),
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'OpenCode Hooks',
               description:
                   'Use an Alera-managed OpenCode config overlay with status plugin.',
@@ -112,7 +131,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
                   .read(settingsControllerProvider.notifier)
                   .setAgentStatusHookEnabled(AgentType.opencode, value),
             ),
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'Pi Hooks',
               description:
                   'Use an Alera-managed Pi agent overlay with status extension.',
@@ -121,7 +140,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
                   .read(settingsControllerProvider.notifier)
                   .setAgentStatusHookEnabled(AgentType.pi, value),
             ),
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'Amp Hooks',
               description: 'Use an Alera-managed Amp config overlay.',
               value: agents.agentStatusHooks.amp,
@@ -129,7 +148,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
                   .read(settingsControllerProvider.notifier)
                   .setAgentStatusHookEnabled(AgentType.amp, value),
             ),
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'Agent Status Notifications',
               description:
                   'Show native notifications when an agent needs attention or finishes.',
@@ -138,7 +157,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
                   .read(settingsControllerProvider.notifier)
                   .setAgentStatusNotificationsEnabled(value),
             ),
-            _SwitchSettingRow(
+            SettingsSwitchRow(
               title: 'Keep Computer Awake While Agents Are Working',
               description: _agentAwakeSettingDescription(
                 Theme.of(context).platform,
@@ -154,123 +173,7 @@ class _GeneralSettingsPane extends ConsumerWidget {
         const UpdateSettingsSection(),
         if (starState != GitHubStarState.hidden) ...<Widget>[
           const SizedBox(height: AleraTokens.space24),
-          _SupportAleraSection(state: starState),
-        ],
-      ],
-    );
-  }
-}
-
-class _AleraCliSkillControl extends ConsumerStatefulWidget {
-  const _AleraCliSkillControl();
-
-  @override
-  ConsumerState<_AleraCliSkillControl> createState() =>
-      _AleraCliSkillControlState();
-}
-
-class _AleraCliSkillControlState extends ConsumerState<_AleraCliSkillControl> {
-  bool _installing = false;
-  String? _status;
-
-  Future<void> _install() async {
-    if (_installing) {
-      return;
-    }
-    setState(() {
-      _installing = true;
-      _status = null;
-    });
-    try {
-      final result = await ref
-          .read(aleraCliSkillServiceProvider)
-          .installOrUpdate();
-      if (!mounted) {
-        return;
-      }
-      setState(() {
-        _status = result.summary;
-      });
-    } catch (error) {
-      if (!mounted) {
-        return;
-      }
-      setState(() {
-        _status = 'Install Failed: $error';
-      });
-    } finally {
-      if (mounted) {
-        setState(() {
-          _installing = false;
-        });
-      }
-    }
-  }
-
-  Future<void> _copyCommand() async {
-    await Clipboard.setData(
-      const ClipboardData(text: aleraCliSkillInstallCommand),
-    );
-    if (!mounted) {
-      return;
-    }
-    setState(() {
-      _status = 'Install Command Copied';
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final status = _status;
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.end,
-      mainAxisSize: MainAxisSize.min,
-      children: <Widget>[
-        Wrap(
-          alignment: WrapAlignment.end,
-          spacing: AleraTokens.space8,
-          runSpacing: AleraTokens.space8,
-          children: <Widget>[
-            SizedBox(
-              height: _kSupportControlHeight,
-              child: OutlinedButton.icon(
-                onPressed: _installing ? null : _copyCommand,
-                icon: const Icon(AleraIcons.copy, size: 16),
-                label: const Text('Copy'),
-              ),
-            ),
-            SizedBox(
-              height: _kSupportControlHeight,
-              child: FilledButton.tonalIcon(
-                onPressed: _installing ? null : _install,
-                icon: _installing
-                    ? const SizedBox(
-                        width: 14,
-                        height: 14,
-                        child: CircularProgressIndicator(
-                          strokeWidth: 2,
-                          color: AleraTokens.foreground,
-                        ),
-                      )
-                    : const Icon(AleraIcons.download, size: 16),
-                label: Text(_installing ? 'Installing' : 'Install / Update'),
-              ),
-            ),
-          ],
-        ),
-        if (status != null) ...<Widget>[
-          const SizedBox(height: AleraTokens.space6),
-          Text(
-            status,
-            textAlign: TextAlign.right,
-            maxLines: 2,
-            overflow: TextOverflow.ellipsis,
-            style: Theme.of(context).textTheme.bodySmall?.copyWith(
-              color: status.startsWith('Install Failed')
-                  ? AleraTokens.error
-                  : AleraTokens.foregroundMuted,
-            ),
-          ),
+          SupportAleraSection(state: starState),
         ],
       ],
     );
@@ -282,195 +185,6 @@ String _agentAwakeSettingDescription(TargetPlatform platform) {
     return 'Keeps this computer and display awake while agents are working. Lid-close behavior follows this device\'s power settings.';
   }
   return 'Keeps this computer and display awake while agents are working. Alera also asks this device to stay awake when the lid is closed, subject to its power policy.';
-}
-
-class _SupportAleraSection extends ConsumerWidget {
-  const _SupportAleraSection({required this.state});
-
-  final GitHubStarState state;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final theme = Theme.of(context);
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: <Widget>[
-        Padding(
-          padding: const EdgeInsets.only(
-            left: AleraTokens.space4,
-            bottom: AleraTokens.space8,
-          ),
-          child: Text(
-            'Support Alera',
-            style: theme.textTheme.titleSmall?.copyWith(
-              color: AleraTokens.foreground,
-              fontWeight: FontWeight.w600,
-            ),
-          ),
-        ),
-        AleraPanel(
-          children: <Widget>[
-            AleraSettingRow(
-              title: 'Star Alera on GitHub',
-              description: null,
-              child: _StarControl(
-                state: state,
-                onStar: () =>
-                    ref.read(gitHubStarControllerProvider.notifier).star(),
-              ),
-            ),
-          ],
-        ),
-      ],
-    );
-  }
-}
-
-class _StarControl extends StatelessWidget {
-  const _StarControl({required this.state, required this.onStar});
-
-  final GitHubStarState state;
-  final VoidCallback onStar;
-
-  @override
-  Widget build(BuildContext context) {
-    return Align(
-      alignment: Alignment.centerRight,
-      child: AnimatedSwitcher(
-        duration: AleraTokens.durationMid,
-        transitionBuilder: (child, animation) {
-          return FadeTransition(
-            opacity: animation,
-            child: SlideTransition(
-              position: Tween<Offset>(
-                begin: const Offset(0.05, 0),
-                end: Offset.zero,
-              ).animate(animation),
-              child: child,
-            ),
-          );
-        },
-        child: _buildChild(),
-      ),
-    );
-  }
-
-  Widget _buildChild() {
-    return switch (state) {
-      GitHubStarState.loading => const _StarSkeleton(
-        key: ValueKey<String>('loading'),
-      ),
-      GitHubStarState.notStarred => _StarButton(
-        key: const ValueKey<String>('not-starred'),
-        label: 'Star',
-        onPressed: onStar,
-      ),
-      GitHubStarState.starring => const _StarButton(
-        key: ValueKey<String>('starring'),
-        label: 'Starring…',
-        busy: true,
-      ),
-      GitHubStarState.starred => const _StarThanks(
-        key: ValueKey<String>('starred'),
-      ),
-      GitHubStarState.error => _StarButton(
-        key: const ValueKey<String>('error'),
-        label: 'Try again',
-        onPressed: onStar,
-      ),
-      GitHubStarState.hidden => const SizedBox.shrink(
-        key: ValueKey<String>('hidden'),
-      ),
-    };
-  }
-}
-
-@visibleForTesting
-Widget buildStarControlForTesting({
-  required GitHubStarState state,
-  required VoidCallback onStar,
-}) {
-  return _StarControl(state: state, onStar: onStar);
-}
-
-class _StarButton extends StatelessWidget {
-  const _StarButton({
-    super.key,
-    required this.label,
-    this.onPressed,
-    this.busy = false,
-  });
-
-  final String label;
-  final VoidCallback? onPressed;
-  final bool busy;
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      height: _kSupportControlHeight,
-      child: FilledButton.tonalIcon(
-        onPressed: onPressed,
-        icon: busy
-            ? const SizedBox(
-                width: 14,
-                height: 14,
-                child: CircularProgressIndicator(
-                  strokeWidth: 2,
-                  color: AleraTokens.foreground,
-                ),
-              )
-            : const Icon(AleraIcons.star, size: 16),
-        label: Text(label),
-      ),
-    );
-  }
-}
-
-class _StarThanks extends StatelessWidget {
-  const _StarThanks({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Semantics(
-      label: 'Thanks for starring Alera',
-      liveRegion: true,
-      child: SizedBox(
-        height: _kSupportControlHeight,
-        child: Row(
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-            const Icon(AleraIcons.star, size: 16, color: AleraTokens.warning),
-            const SizedBox(width: AleraTokens.space6),
-            Text(
-              'Thanks for the support!',
-              style: theme.textTheme.bodyMedium?.copyWith(
-                color: AleraTokens.warning,
-                fontWeight: FontWeight.w500,
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _StarSkeleton extends StatelessWidget {
-  const _StarSkeleton({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: _kSupportControlHeight,
-      width: 110,
-      decoration: BoxDecoration(
-        color: AleraTokens.surfaceElevated,
-        borderRadius: BorderRadius.circular(AleraTokens.radiusLg),
-      ),
-    );
-  }
 }
 
 class _WorkspaceDirectoryRow extends StatefulWidget {
