+++ lib/src/features/settings/presentation/settings_dialog.dart
@@ -1,60 +1,24 @@
-import 'dart:async';
-
 import 'package:alera/src/app/providers.dart';
 import 'package:alera/src/app/theme/alera_tokens.dart';
 import 'package:alera/src/design_system/icons/alera_icons.dart';
-import 'package:alera/src/features/ai_text_generation/application/ai_text_generation_providers.dart';
-import 'package:alera/src/features/ai_text_generation/application/ai_text_generation_registry.dart';
-import 'package:alera/src/features/ai_text_generation/application/ai_text_model_discovery_service.dart';
-import 'package:alera/src/features/ai_text_generation/domain/ai_text_generation_settings.dart';
-import 'package:alera/src/features/projects/domain/project.dart';
-import 'package:alera/src/features/projects/domain/project_config.dart';
-import 'package:alera/src/features/projects/domain/project_config_paths.dart';
-import 'package:alera/src/features/remote_hosts/application/ssh_target_providers.dart';
-import 'package:alera/src/features/remote_hosts/domain/ssh_target.dart';
-import 'package:alera/src/design_system/buttons/alera_icon_button.dart';
-import 'package:alera/src/design_system/buttons/alera_segmented_button.dart';
-import 'package:alera/src/design_system/feedback/alera_color_swatch.dart';
-import 'package:alera/src/design_system/feedback/alera_empty_state.dart';
-import 'package:alera/src/design_system/forms/alera_number_field.dart';
-import 'package:alera/src/design_system/forms/alera_search_field.dart';
-import 'package:alera/src/design_system/forms/alera_setting_row.dart';
-import 'package:alera/src/design_system/forms/alera_text_field.dart';
 import 'package:alera/src/design_system/layout/alera_dialog.dart';
-import 'package:alera/src/design_system/menus/alera_dropdown_entry.dart';
-import 'package:alera/src/design_system/menus/alera_menu_item.dart';
-import 'package:alera/src/design_system/surfaces/alera_panel.dart';
-import 'package:alera/src/features/agent_status/domain/agent_status.dart';
 import 'package:alera/src/features/keyboard/presentation/keyboard_settings_pane.dart';
-import 'package:alera/src/features/settings/domain/alera_settings.dart';
-import 'package:alera/src/features/settings/domain/editor_syntax_theme_catalog.dart';
-import 'package:alera/src/features/settings/domain/terminal_theme_catalog.dart';
-import 'package:alera/src/features/settings/infra/alera_cli_skill_service.dart';
 import 'package:alera/src/features/settings/infra/system_font_service.dart';
-import 'package:alera/src/features/updater/presentation/update_settings_section.dart';
-import 'package:file_selector/file_selector.dart';
+import 'package:alera/src/features/settings/presentation/panes/ai_text_pane.dart';
+import 'package:alera/src/features/settings/presentation/panes/editor_pane.dart';
+import 'package:alera/src/features/settings/presentation/panes/general_pane.dart';
+import 'package:alera/src/features/settings/presentation/panes/projects_pane.dart';
+import 'package:alera/src/features/settings/presentation/panes/remote_hosts_pane.dart';
+import 'package:alera/src/features/settings/presentation/panes/terminal_pane.dart';
+import 'package:alera/src/features/settings/presentation/settings_dialog_content.dart';
+import 'package:alera/src/features/settings/presentation/settings_dialog_sidebar.dart';
+import 'package:alera/src/features/settings/presentation/settings_search_entries.dart';
+import 'package:alera/src/features/settings/presentation/settings_sections.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter/services.dart';
-
-part 'settings_dialog_navigation.dart';
-part 'settings_dialog_general_pane.dart';
-part 'settings_dialog_projects_pane.dart';
-part 'settings_dialog_remote_hosts_pane.dart';
-part 'settings_dialog_ai_text_pane.dart';
-part 'settings_dialog_editor_pane.dart';
-part 'settings_dialog_terminal_pane.dart';
-part 'settings_dialog_theme_picker.dart';
-part 'settings_dialog_setting_rows.dart';
-part 'settings_dialog_theme_controls.dart';
 
 const double _kDialogMaxWidth = 920;
 const double _kDialogMaxHeight = 680;
-const double _kSidebarWidth = 260;
-const double _kSidebarIconSize = 16;
-const double _kSectionIconSize = 18;
-const double _kSupportControlHeight = 34;
-const double _kPickerMenuMaxHeight = 220;
 
 class SettingsDialog extends ConsumerStatefulWidget {
   const SettingsDialog({super.key});
@@ -102,77 +66,77 @@ class _SettingsDialogState extends ConsumerState<SettingsDialog> {
     final settings = ref.watch(settingsControllerProvider);
     final controller = ref.read(settingsControllerProvider.notifier);
 
-    final sections = <_SettingsSectionData>[
-      _SettingsSectionData(
+    final sections = <SettingsSectionData>[
+      SettingsSectionData(
         id: 'general',
         title: 'General',
         description: 'Storage and integrations.',
         icon: AleraIcons.tune,
-        entries: _generalSearchEntries,
-        builder: (_) => _GeneralSettingsPane(
+        entries: generalSearchEntries,
+        builder: (_) => GeneralSettingsPane(
           general: settings.general,
           agents: settings.agents,
         ),
       ),
-      _SettingsSectionData(
+      SettingsSectionData(
         id: 'projects',
         title: 'Projects',
         description: 'Per-Project Workspace Setup.',
         icon: AleraIcons.folderSpecial,
-        entries: _projectSearchEntries,
-        builder: (_) => const _ProjectSettingsPane(),
+        entries: projectSearchEntries,
+        builder: (_) => const ProjectSettingsPane(),
       ),
-      _SettingsSectionData(
+      SettingsSectionData(
         id: 'remoteHosts',
         title: 'Remote Hosts',
         description: 'SSH runtime targets.',
         icon: AleraIcons.host,
-        entries: _remoteHostSearchEntries,
-        builder: (_) => const _RemoteHostSettingsPane(),
+        entries: remoteHostSearchEntries,
+        builder: (_) => const RemoteHostSettingsPane(),
       ),
-      _SettingsSectionData(
+      SettingsSectionData(
         id: 'editor',
         title: 'Editor',
         description: 'Code editor defaults.',
         icon: AleraIcons.code,
-        entries: _editorSearchEntries,
+        entries: editorSearchEntries,
         onReset: controller.resetEditorSettings,
-        builder: (_) => _EditorSettingsPane(
+        builder: (_) => EditorSettingsPane(
           settings: settings.editor,
           onChanged: (editor) => controller.updateEditor(editor),
         ),
       ),
-      _SettingsSectionData(
+      SettingsSectionData(
         id: 'aiText',
         title: 'AI Text',
         description: 'AI-generated source control text.',
         icon: AleraIcons.ai,
-        entries: _aiTextSearchEntries,
+        entries: aiTextSearchEntries,
         onReset: controller.resetAiTextGenerationSettings,
-        builder: (_) => _AiTextSettingsPane(
+        builder: (_) => AiTextSettingsPane(
           settings: settings.aiTextGeneration,
           onChanged: (aiText) => controller.updateAiTextGeneration(aiText),
         ),
       ),
-      _SettingsSectionData(
+      SettingsSectionData(
         id: 'terminal',
         title: 'Terminal',
         description: 'Appearance defaults for new terminal sessions.',
         icon: AleraIcons.terminal,
-        entries: _terminalSearchEntries,
+        entries: terminalSearchEntries,
         onReset: controller.resetTerminalSettings,
-        builder: (_) => _TerminalSettingsPane(
+        builder: (_) => TerminalSettingsPane(
           settings: settings.terminal,
           fontSuggestions: _fontSuggestions,
           onChanged: (terminal) => controller.updateTerminal(terminal),
         ),
       ),
-      _SettingsSectionData(
+      SettingsSectionData(
         id: 'keyboard',
         title: 'Keyboard',
         description: 'Shortcuts and key bindings.',
         icon: AleraIcons.keyboard,
-        entries: _keyboardSearchEntries,
+        entries: keyboardSearchEntries,
         onReset: controller.resetKeyboardShortcuts,
         builder: (_) => const KeyboardSettingsPane(),
       ),
@@ -195,7 +159,7 @@ class _SettingsDialogState extends ConsumerState<SettingsDialog> {
       child: Row(
         crossAxisAlignment: CrossAxisAlignment.stretch,
         children: <Widget>[
-          _SettingsSidebar(
+          SettingsSidebar(
             queryController: _searchController,
             visibleSections: visibleSections,
             activeSectionId: activeSection?.id,
@@ -204,365 +168,17 @@ class _SettingsDialogState extends ConsumerState<SettingsDialog> {
           const VerticalDivider(width: 1, color: AleraTokens.borderSubtle),
           Expanded(
             child: activeSection != null
-                ? _SettingsContent(
+                ? SettingsContent(
                     section: activeSection,
                     onClose: () => Navigator.of(context).pop(),
                   )
-                : _NoSettingsResults(
-                    onClose: () => Navigator.of(context).pop(),
-                  ),
+                : NoSettingsResults(onClose: () => Navigator.of(context).pop()),
           ),
         ],
       ),
     );
   }
 }
-
-class _SettingsSectionData {
-  const _SettingsSectionData({
-    required this.id,
-    required this.title,
-    required this.description,
-    required this.icon,
-    required this.entries,
-    required this.builder,
-    this.onReset,
-  });
-
-  final String id;
-  final String title;
-  final String description;
-  final IconData icon;
-  final List<_SettingsSearchEntry> entries;
-  final WidgetBuilder builder;
-  final Future<void> Function()? onReset;
-
-  bool matches(String query) {
-    if (query.isEmpty) {
-      return true;
-    }
-    final sectionEntry = _SettingsSearchEntry(
-      title: title,
-      description: description,
-    );
-    return sectionEntry.matches(query) ||
-        entries.any((entry) => entry.matches(query));
-  }
-}
-
-class _SettingsSearchEntry {
-  const _SettingsSearchEntry({
-    required this.title,
-    this.description,
-    this.keywords = const <String>[],
-  });
-
-  final String title;
-  final String? description;
-  final List<String> keywords;
-
-  bool matches(String query) {
-    if (query.isEmpty) {
-      return true;
-    }
-    return <String>[
-      title,
-      description ?? '',
-      ...keywords,
-    ].any((value) => value.toLowerCase().contains(query));
-  }
-}
-
-const List<_SettingsSearchEntry> _generalSearchEntries = <_SettingsSearchEntry>[
-  _SettingsSearchEntry(
-    title: 'Workspace Directory',
-    description: 'Where new linked workspaces are created on disk.',
-    keywords: <String>['worktree', 'folder', 'location', 'path'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Confirm Project Removal',
-    description: 'Ask before unregistering a project.',
-    keywords: <String>['safety', 'destructive', 'remove', 'delete'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Confirm Workspace Removal',
-    description: 'Ask before removing a workspace worktree.',
-    keywords: <String>['safety', 'destructive', 'remove', 'delete'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Alera CLI Skill',
-    description: 'Install agent instructions for the Alera CLI.',
-    keywords: <String>['codex', 'skill', 'cli', 'agent', 'workspace'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Codex Hooks',
-    description: 'Use Alera-managed Codex runtime hooks.',
-    keywords: <String>['codex', 'agent', 'status', 'hooks'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Claude Code Hooks',
-    description: 'Use an Alera-managed Claude Code config with status hooks.',
-    keywords: <String>['claude', 'agent', 'status', 'hooks'],
-  ),
-  _SettingsSearchEntry(
-    title: 'GitHub Copilot Hooks',
-    description: 'Use an Alera-managed GitHub Copilot home overlay.',
-    keywords: <String>['copilot', 'github', 'agent', 'status', 'hooks'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Cursor Hooks',
-    description: 'Use an Alera-managed Cursor Agent plugin wrapper.',
-    keywords: <String>['cursor', 'agent', 'status', 'hooks', 'cli'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Antigravity Hooks',
-    description: 'Install managed Antigravity hooks for the agy CLI.',
-    keywords: <String>['antigravity', 'agy', 'agent', 'status', 'hooks'],
-  ),
-  _SettingsSearchEntry(
-    title: 'OpenCode Hooks',
-    description: 'Install managed OpenCode status plugin.',
-    keywords: <String>['opencode', 'agent', 'status', 'hooks', 'plugin'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Pi Hooks',
-    description: 'Install managed Pi status extension.',
-    keywords: <String>['pi', 'agent', 'status', 'hooks', 'extension'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Amp Hooks',
-    description: 'Use an Alera-managed Amp config overlay.',
-    keywords: <String>['amp', 'agent', 'status', 'hooks', 'plugin'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Agent Status Notifications',
-    description: 'Show native notifications when agents need attention.',
-    keywords: <String>[
-      'codex',
-      'claude',
-      'copilot',
-      'cursor',
-      'antigravity',
-      'agy',
-      'opencode',
-      'pi',
-      'amp',
-      'agent',
-      'status',
-      'notification',
-    ],
-  ),
-  _SettingsSearchEntry(
-    title: 'Keep Computer Awake While Agents Are Working',
-    description: 'Keep this computer and display awake during agent work.',
-    keywords: <String>[
-      'awake',
-      'sleep',
-      'power',
-      'agent',
-      'working',
-      'lid',
-      'display',
-    ],
-  ),
-  _SettingsSearchEntry(
-    title: 'Updates',
-    description: 'Check desktop releases for this platform.',
-    keywords: <String>['release', 'download', 'version'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Star Alera on GitHub',
-    description: 'Show your support for the project.',
-    keywords: <String>['support', 'github', 'star'],
-  ),
-];
-
-const List<_SettingsSearchEntry> _projectSearchEntries = <_SettingsSearchEntry>[
-  _SettingsSearchEntry(
-    title: 'Project Worktree Setup',
-    description: 'Copy Files And Run Setup Commands For Linked Workspaces.',
-    keywords: <String>[
-      'project',
-      'repo',
-      'worktree',
-      'workspace',
-      'copy',
-      'setup',
-      'alera.toml',
-    ],
-  ),
-];
-
-const List<_SettingsSearchEntry> _remoteHostSearchEntries =
-    <_SettingsSearchEntry>[
-      _SettingsSearchEntry(
-        title: 'Remote Hosts',
-        description: 'Manage SSH targets and remote runtime bootstrap.',
-        keywords: <String>[
-          'ssh',
-          'remote',
-          'host',
-          'runtime',
-          'bootstrap',
-          'mobile',
-        ],
-      ),
-    ];
-
-const List<_SettingsSearchEntry> _keyboardSearchEntries =
-    <_SettingsSearchEntry>[
-      _SettingsSearchEntry(
-        title: 'Keyboard Shortcuts',
-        description: 'View and remap app-wide key bindings.',
-        keywords: <String>[
-          'shortcut',
-          'hotkey',
-          'keybinding',
-          'binding',
-          'keymap',
-        ],
-      ),
-      _SettingsSearchEntry(
-        title: 'Terminal Shortcut Behavior',
-        description:
-            'Choose whether app shortcuts win while a terminal is '
-            'focused.',
-        keywords: <String>['app first', 'terminal first', 'policy'],
-      ),
-    ];
-
-const List<_SettingsSearchEntry> _editorSearchEntries = <_SettingsSearchEntry>[
-  _SettingsSearchEntry(
-    title: 'Theme Preset',
-    description: 'Syntax highlighting theme used by editor tabs.',
-    keywords: <String>['syntax', 'highlight', 'highlighting', 'color', 'code'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Tab Size',
-    description: 'Spaces inserted when pressing Tab in editor tabs.',
-    keywords: <String>['indent', 'indentation', 'spaces', 'code'],
-  ),
-];
-
-const List<_SettingsSearchEntry> _aiTextSearchEntries = <_SettingsSearchEntry>[
-  _SettingsSearchEntry(
-    title: 'AI Text Generation',
-    description: 'Generate source control text with local agent CLIs.',
-    keywords: <String>['ai', 'commit', 'pull request', 'branch'],
-  ),
-  _SettingsSearchEntry(
-    title: 'AI Text Agent',
-    description: 'Choose the CLI used for generated text.',
-    keywords: <String>[
-      'codex',
-      'claude',
-      'copilot',
-      'cursor',
-      'antigravity',
-      'agy',
-      'opencode',
-      'pi',
-      'amp',
-      'custom',
-    ],
-  ),
-  _SettingsSearchEntry(
-    title: 'AI Text Instructions',
-    description: 'Commit message instructions for generated text.',
-    keywords: <String>['prompt', 'instructions', 'commit message'],
-  ),
-];
-
-const List<_SettingsSearchEntry>
-_terminalSearchEntries = <_SettingsSearchEntry>[
-  _SettingsSearchEntry(
-    title: 'Font Family',
-    description: 'Typeface used in new terminal sessions.',
-    keywords: <String>['monospace', 'jetbrains', 'typeface'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Font Size',
-    description: 'Text size used in new terminal sessions.',
-    keywords: <String>['terminal text', 'zoom'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Font Weight',
-    description: 'Weight used for terminal text.',
-    keywords: <String>['terminal text', 'bold'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Line Height',
-    description: 'Vertical spacing for terminal rows.',
-    keywords: <String>['spacing', 'rows'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Theme Preset',
-    description: 'Built-in terminal color theme.',
-    keywords: <String>['color', 'appearance', 'palette'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Background Opacity',
-    description: 'Opacity of the terminal background.',
-    keywords: <String>['transparent', 'alpha'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Horizontal Padding',
-    description: 'Horizontal spacing around the terminal grid.',
-    keywords: <String>['inset', 'space'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Vertical Padding',
-    description: 'Vertical spacing around the terminal grid.',
-    keywords: <String>['inset', 'space'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Cursor Shape',
-    description: 'Cursor style for new terminal sessions.',
-    keywords: <String>['caret', 'block', 'bar', 'underline'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Blinking Cursor',
-    description: 'Blink the terminal cursor while focused.',
-    keywords: <String>['caret', 'blink'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Cursor Opacity',
-    description: 'Opacity of the terminal cursor.',
-    keywords: <String>['caret', 'alpha'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Color Overrides',
-    description: 'Override core terminal colors.',
-    keywords: <String>['foreground', 'background', 'selection', 'cursor'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Scrollback Lines',
-    description: 'Maximum terminal history retained per session.',
-    keywords: <String>['history', 'buffer'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Host Scrollback Size',
-    description: 'Maximum host-side terminal output retained per session.',
-    keywords: <String>['history', 'buffer', 'memory', 'host'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Empty Host Shutdown',
-    description:
-        'Stop the terminal host after the app closes with no sessions.',
-    keywords: <String>['host', 'sidecar', 'lifetime', 'timeout'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Detached Session Shutdown',
-    description:
-        'Stop detached running terminal sessions after the app stays closed.',
-    keywords: <String>['host', 'sidecar', 'session', 'timeout'],
-  ),
-  _SettingsSearchEntry(
-    title: 'Word Separators',
-    description: 'Characters that break double-click word selection.',
-    keywords: <String>['boundary', 'selection', 'double click'],
-  ),
-];
 
 List<String> _mergeFontSuggestions(List<String> fonts) {
   final byName = <String, String>{};
@@ -575,20 +191,4 @@ List<String> _mergeFontSuggestions(List<String> fonts) {
   }
   return byName.values.toList()
     ..sort((a, b) => a.toLowerCase().compareTo(b.toLowerCase()));
-}
-
-List<String> _filterOrdered(Iterable<String> values, String query) {
-  final normalized = query.trim().toLowerCase();
-  final matches = normalized.isEmpty
-      ? values.toList(growable: false)
-      : <String>[
-          ...values.where(
-            (value) => value.toLowerCase().startsWith(normalized),
-          ),
-          ...values.where((value) {
-            final lower = value.toLowerCase();
-            return !lower.startsWith(normalized) && lower.contains(normalized);
-          }),
-        ];
-  return matches;
 }
