+++ lib/src/features/settings/presentation/settings_dialog_sidebar.dart
@@ -1,7 +1,15 @@
-part of 'settings_dialog.dart';
+import 'package:alera/src/app/theme/alera_tokens.dart';
+import 'package:alera/src/design_system/feedback/alera_empty_state.dart';
+import 'package:alera/src/design_system/forms/alera_search_field.dart';
+import 'package:alera/src/features/settings/presentation/settings_sections.dart';
+import 'package:flutter/material.dart';
 
-class _SettingsSidebar extends StatelessWidget {
-  const _SettingsSidebar({
+const double _kSidebarWidth = 260;
+const double _kSidebarIconSize = 16;
+
+class SettingsSidebar extends StatelessWidget {
+  const SettingsSidebar({
+    super.key,
     required this.queryController,
     required this.visibleSections,
     required this.activeSectionId,
@@ -9,7 +17,7 @@ class _SettingsSidebar extends StatelessWidget {
   });
 
   final TextEditingController queryController;
-  final List<_SettingsSectionData> visibleSections;
+  final List<SettingsSectionData> visibleSections;
   final String? activeSectionId;
   final ValueChanged<String> onSelect;
 
@@ -60,7 +68,7 @@ class _SettingsSidebar extends StatelessWidget {
                           const SizedBox(height: AleraTokens.space2),
                       itemBuilder: (_, index) {
                         final section = visibleSections[index];
-                        return _SettingsNavItem(
+                        return SettingsNavItem(
                           section: section,
                           active: section.id == activeSectionId,
                           onTap: () => onSelect(section.id),
@@ -75,14 +83,15 @@ class _SettingsSidebar extends StatelessWidget {
   }
 }
 
-class _SettingsNavItem extends StatelessWidget {
-  const _SettingsNavItem({
+class SettingsNavItem extends StatelessWidget {
+  const SettingsNavItem({
+    super.key,
     required this.section,
     required this.active,
     required this.onTap,
   });
 
-  final _SettingsSectionData section;
+  final SettingsSectionData section;
   final bool active;
   final VoidCallback onTap;
 
@@ -130,69 +139,6 @@ class _SettingsNavItem extends StatelessWidget {
           ),
         ),
       ),
-    );
-  }
-}
-
-class _SettingsContent extends StatelessWidget {
-  const _SettingsContent({required this.section, required this.onClose});
-
-  final _SettingsSectionData section;
-  final VoidCallback onClose;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return ListView(
-      padding: const EdgeInsets.all(AleraTokens.space24),
-      children: <Widget>[
-        Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-            Row(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: <Widget>[
-                Icon(
-                  section.icon,
-                  size: _kSectionIconSize,
-                  color: AleraTokens.accent,
-                ),
-                const SizedBox(width: AleraTokens.space8),
-                Expanded(
-                  child: Text(section.title, style: theme.textTheme.titleLarge),
-                ),
-                if (section.onReset != null) ...<Widget>[
-                  const SizedBox(width: AleraTokens.space8),
-                  TextButton(
-                    onPressed: () async {
-                      FocusManager.instance.primaryFocus?.unfocus();
-                      await Future<void>.delayed(Duration.zero);
-                      await section.onReset!();
-                    },
-                    child: Text('Reset ${section.title}'),
-                  ),
-                ],
-                const SizedBox(width: AleraTokens.space4),
-                AleraIconButton(
-                  tooltip: 'Close',
-                  onPressed: onClose,
-                  icon: AleraIcons.close,
-                  minSize: 34,
-                ),
-              ],
-            ),
-            const SizedBox(height: AleraTokens.space4),
-            Text(
-              section.description,
-              style: theme.textTheme.bodySmall?.copyWith(
-                color: AleraTokens.foregroundMuted,
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: AleraTokens.space20),
-        section.builder(context),
-      ],
     );
   }
 }
